bot/requirements.txt
python-dotenv==1.0.0
Pyrogram==2.0.106
TgCrypto==1.2.5
motor==3.3.0
pydantic==2.3.0
aio-pika==9.2.2